fix(imeta): accept floating-point dimensions so image space is reserved pre-load
Primal-style clients emit "dim 317.0x498.0" in NIP-92 imeta tags. DimensionTag.parse
called Int.parseInt on each component, threw NumberFormatException, and returned
null. With dim==null and no cached aspect ratio, GifVideoView / UrlImageView built
the container without an aspectRatio modifier, so the inline image collapsed to
zero height and the post body looked empty until Coil delivered the bitmap.

Parse each component as Double then truncate to Int. Adds DimensionTagTest covering
integer, float, truncation, 0x0 and malformed inputs.

+++ quartz/src/commonMain/kotlin/com/vitorpamplona/quartz/nip94FileMetadata/tags/DimensionTag.kt
@@ -56,8 +56,11 @@ class DimensionTag(
             if (parts.size != 2) return null
 
             return try {
-                val width = parts[0].toInt()
-                val height = parts[1].toInt()
+                // Some clients (e.g. Primal) emit floating-point dimensions like "317.0x498.0"
+                // in NIP-92 imeta tags. Parse as Double and truncate to keep those tags usable
+                // for pre-load layout reservation.
+                val width = parts[0].toDouble().toInt()
+                val height = parts[1].toDouble().toInt()
 
                 DimensionTag(width, height)
             } catch (e: Exception) {
